adds MEP submodule structure
